+++ src/structure.rs
@@ -89,7 +89,7 @@ impl Structure {
                         r#" AND "_id" IN({sql_params})"#,
                         sql_params = sql_param_gen.get_as_list(rows.len()),
                     ));
-                    params.extend(rows.iter().map(|row| ParamValue::from(**row)));
+                    params.extend(rows.iter().map(|row| ParamValue::from(**row as i32)));
                 }
                 sql.push_str(r#" RETURNING 1 AS "inserted""#);
                 let rows = rltbl.pool.query(&sql, params).await?;
